+++ lib/flutter_lua.dart
@@ -10,7 +10,6 @@ import 'dart:async' show Future;
 import 'dart:io' show File;
 
 import 'package:flutter/services.dart' show MethodChannel, PlatformException;
-import 'package:meta/meta.dart' show experimental;
 
 import 'src/errors.dart' show LuaError;
 
@@ -22,20 +21,29 @@ abstract class Lua {
     return await _channel.invokeMethod('getVersion');
   }
 
-  @experimental
-  static Future<void> doFile(final File path) async {
+  /// Evaluates a Lua code snippet, returning a result.
+  static Future<dynamic> eval(final String code) async {
     try {
-      final absolutePath = path.isAbsolute ? path : path.absolute;
-      return await _channel.invokeMethod('doFile', absolutePath.toString());
+      return await _channel.invokeMethod('evalString', code);
     } on PlatformException catch (error) {
       throw LuaError.from(error);
     }
   }
 
-  @experimental
-  static Future<void> doString(final String code) async {
+  /// Evaluates a bundled Lua source file, returning a result.
+  static Future<dynamic> evalAsset(final String assetName) async {
     try {
-      return await _channel.invokeMethod('doString', code);
+      return await _channel.invokeMethod('evalAsset', assetName);
+    } on PlatformException catch (error) {
+      throw LuaError.from(error);
+    }
+  }
+
+  /// Evaluates a Lua source file, returning a result.
+  static Future<dynamic> evalFile(final File path) async {
+    try {
+      final absolutePath = path.isAbsolute ? path : path.absolute;
+      return await _channel.invokeMethod('evalFile', absolutePath.toString());
     } on PlatformException catch (error) {
       throw LuaError.from(error);
     }